+++ spec/mock_rack_app.rb
@@ -1,14 +1,11 @@
 class MockRackApp
 
-  attr_reader :request_body
-
   def initialize
     @request_headers = {}
   end
 
   def call(env)
     @env = env
-    @request_body = env['rack.input'].read
     [200, {'Content-Type' => 'text/plain'}, ['OK']]
   end